refactor: 사용하지 않는 주석 제거, 코드 중첩 depth 간소화, 컴포넌트 key index 사용 한 부분 변경
