+++ packages/client/src/components/live/PinnedCourses.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import Tooltip from '@/components/common/Tooltip.tsx';
 import AlarmOptionModal from '@/components/toast/AlarmOptionModal.tsx';
-import RealtimeCard from '@/components/subjectTable/RealtimeCard.tsx';
-import NetworkError from '@/components/dashboard/errors/NetworkError.tsx';
-import ZeroPinError from '@/components/dashboard/errors/ZeroPinError.tsx';
+import RealtimeCard from '@/components/live/subjectTable/RealtimeCard.tsx';
+import NetworkError from '@/components/live/errors/NetworkError.tsx';
+import ZeroPinError from '@/components/live/errors/ZeroPinError.tsx';
 import { usePinned } from '@/store/usePinned.ts';
 import useFindWishes from '@/hooks/useFindWishes.ts';
 import { SSEType, useSseData } from '@/hooks/useSSEManager.ts';
@@ -11,29 +11,11 @@ import useNotification from '@/hooks/useNotification.ts';
 import AlarmBlueIcon from '@/assets/alarm-blue.svg?react';
 import AlarmDisabledIcon from '@/assets/alarm-disabled.svg?react';
 import SettingSvg from '@/assets/settings.svg?react';
-import AlarmButton from './AlarmButton';
+import AlarmButton from './AlarmButton.tsx';
 
 const PinnedCourses = () => {
-  const { data, isPending, isError, refetch } = usePinned();
-  const { data: pinnedSeats } = useSseData(SSEType.PINNED);
-  const pinnedWishes = useFindWishes(data?.map(pinned => pinned.subjectId) ?? []);
   const [isAlarmSettingOpen, setIsAlarmSettingOpen] = useState(false);
-
   const { isAlarm, changeAlarm } = useNotification();
-
-  const getSeats = (subjectId: number) => {
-    if (!pinnedSeats) return -1;
-
-    const pinned = pinnedSeats?.find(pinnedSeat => pinnedSeat.subjectId === subjectId);
-    return pinned?.seatCount ?? -1;
-  };
-
-  const getQueryTime = (subjectId: number) => {
-    if (!pinnedSeats) return '';
-
-    const pinned = pinnedSeats?.find(pinnedSeat => pinnedSeat.subjectId === subjectId);
-    return pinned?.queryTime ?? '';
-  };
 
   return (
     <>
@@ -69,37 +51,59 @@ const PinnedCourses = () => {
           </div>
         </div>
 
-        {/* <p className="text-sm text-gray-500 mb-2">
-          검색중 표시는 최대 5분까지 나올 수 있어요. <br />
-          만약 검색중이 계속 사라지지 않는다면 새로고침을 해주세요.
-        </p> */}
-
-        {isPending ? (
-          <div className="animate-pulse grid grid-cols-1 md:grid-cols-3 gap-4">
-            {[0, 0, 0].map((_, idx) => (
-              <div key={idx} className="bg-gray-300 shadow-sm rounded-lg p-4 h-24" />
-            ))}
-          </div>
-        ) : isError ? (
-          <NetworkError onReload={refetch} />
-        ) : !pinnedWishes || pinnedWishes.length === 0 ? (
-          <ZeroPinError />
-        ) : (
-          <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
-            {pinnedWishes.map(subject => (
-              <RealtimeCard
-                key={`${subject.subjectId}_${subject.subjectCode}_${subject.professorName}`}
-                subject={subject}
-                seats={getSeats(subject.subjectId)}
-                queryTime={getQueryTime(subject.subjectId)}
-              />
-            ))}
-            {pinnedWishes.length < 5 && <AlarmButton />}
-          </div>
-        )}
+        <CoursesArea />
       </div>
     </>
   );
 };
+
+function CoursesArea() {
+  const { data, isPending, isError, refetch } = usePinned();
+  const { data: pinnedSeats } = useSseData(SSEType.PINNED);
+
+  const pinnedWishes = useFindWishes(data?.map(pinned => pinned.subjectId) ?? []);
+
+  const getSeats = (subjectId: number) => {
+    const pinned = pinnedSeats?.find(pinnedSeat => pinnedSeat.subjectId === subjectId);
+    return pinned?.seatCount ?? -1;
+  };
+
+  const getQueryTime = (subjectId: number) => {
+    const pinned = pinnedSeats?.find(pinnedSeat => pinnedSeat.subjectId === subjectId);
+    return pinned?.queryTime ?? '';
+  };
+
+  if (isPending) {
+    return (
+      <div className="animate-pulse grid grid-cols-1 md:grid-cols-3 gap-4">
+        {[0, 0, 0].map((_, idx) => (
+          <div key={'skeleton-card-' + idx} className="bg-gray-300 shadow-sm rounded-lg p-4 h-24" />
+        ))}
+      </div>
+    );
+  }
+
+  if (isError) {
+    return <NetworkError onReload={refetch} />;
+  }
+
+  if (!pinnedWishes || pinnedWishes.length === 0) {
+    return <ZeroPinError />;
+  }
+
+  return (
+    <div className="grid grid-cols-1 md:grid-cols-3 gap-4">
+      {pinnedWishes.map(subject => (
+        <RealtimeCard
+          key={`${subject.subjectId}_${subject.subjectCode}_${subject.professorName}`}
+          subject={subject}
+          seats={getSeats(subject.subjectId)}
+          queryTime={getQueryTime(subject.subjectId)}
+        />
+      ))}
+      {pinnedWishes.length < 5 && <AlarmButton />}
+    </div>
+  );
+}
 
 export default PinnedCourses;
