+++ packages/client/src/components/live/errors/NetworkError.tsx
@@ -1,4 +1,4 @@
-import NoneLayout from '@/components/dashboard/NoneLayout.tsx';
+import NoneLayout from '@/components/live/NoneLayout.tsx';
 import ImportantSVG from '@/assets/important.svg?react';
 import ReloadSvg from '@/assets/reload-white.svg?react';
 
@@ -6,10 +6,10 @@ interface NetworkErrorProps {
   onReload: () => void;
 }
 
-function NetworkError({ onReload }: NetworkErrorProps) {
+function NetworkError({ onReload }: Readonly<NetworkErrorProps>) {
   return (
     <NoneLayout
-      title="핀 고정된 과목을 불러올 수 없습니다"
+      title="알림 과목을 불러올 수 없습니다"
       description="네트워크 연결을 확인해주세요"
       icon={<ImportantSVG className="w-7 h-7" />}
     >
